mini cleanup
