+++ qupid/__init__.py
@@ -1,7 +1,9 @@
 from .casematch import (CaseMatchOneToMany, CaseMatchOneToOne,
-                        match_by_single, match_by_multiple)
+                        CaseMatchCollection)
+from .qupid import match_by_single, match_by_multiple, shuffle
+
 
 __version__ = "0.0.2"
 
-__all__ = ["CaseMatchOneToMany", "CaseMatchOneToOne",
-           "match_by_single", "match_by_multiple"]
+__all__ = ["CaseMatchOneToMany", "CaseMatchOneToOne", "match_by_single",
+           "match_by_multiple", "shuffle"]
